chore : 리뷰 반영(메소드 위치 변경 및 Response userType 삭제)
