+++ src/main/java/in/koreatech/koin/domain/user/controller/UserApi.java
@@ -13,14 +13,12 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 
 import in.koreatech.koin.domain.user.dto.AuthResponse;
-import in.koreatech.koin.domain.user.dto.CoopLoginRequest;
-import in.koreatech.koin.domain.user.dto.CoopLoginResponse;
 import in.koreatech.koin.domain.user.dto.CoopResponse;
 import in.koreatech.koin.domain.user.dto.EmailCheckExistsRequest;
 import in.koreatech.koin.domain.user.dto.FindPasswordRequest;
 import in.koreatech.koin.domain.user.dto.NicknameCheckExistsRequest;
-import in.koreatech.koin.domain.user.dto.OwnerLoginRequest;
-import in.koreatech.koin.domain.user.dto.OwnerLoginResponse;
+import in.koreatech.koin.domain.owner.dto.OwnerLoginRequest;
+import in.koreatech.koin.domain.owner.dto.OwnerLoginResponse;
 import in.koreatech.koin.domain.user.dto.StudentLoginRequest;
 import in.koreatech.koin.domain.user.dto.StudentLoginResponse;
 import in.koreatech.koin.domain.user.dto.StudentRegisterRequest;
@@ -119,34 +117,6 @@ public interface UserApi {
     @PostMapping("/student/login")
     ResponseEntity<StudentLoginResponse> studentLogin(
         @RequestBody @Valid StudentLoginRequest request
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "201"),
-            @ApiResponse(responseCode = "400", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
-        }
-    )
-    @Operation(summary = "사장님 로그인")
-    @PostMapping("/owner/login")
-    ResponseEntity<OwnerLoginResponse> ownerLogin(
-        @RequestBody @Valid OwnerLoginRequest request
-    );
-
-    @ApiResponses(
-        value = {
-            @ApiResponse(responseCode = "201"),
-            @ApiResponse(responseCode = "400", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "403", content = @Content(schema = @Schema(hidden = true))),
-            @ApiResponse(responseCode = "404", content = @Content(schema = @Schema(hidden = true))),
-        }
-    )
-    @Operation(summary = "영양사 로그인")
-    @PostMapping("/coop/login")
-    ResponseEntity<CoopLoginResponse> coopLogin(
-        @RequestBody @Valid CoopLoginRequest request
     );
 
     @ApiResponses(
