+++ src/main/java/in/koreatech/koin/domain/user/controller/UserController.java
@@ -20,14 +20,12 @@ import org.springframework.web.servlet.ModelAndView;
 
 import in.koreatech.koin.domain.user.dto.AuthResponse;
 import in.koreatech.koin.domain.user.dto.AuthTokenRequest;
-import in.koreatech.koin.domain.user.dto.CoopLoginRequest;
-import in.koreatech.koin.domain.user.dto.CoopLoginResponse;
 import in.koreatech.koin.domain.user.dto.CoopResponse;
 import in.koreatech.koin.domain.user.dto.EmailCheckExistsRequest;
 import in.koreatech.koin.domain.user.dto.FindPasswordRequest;
 import in.koreatech.koin.domain.user.dto.NicknameCheckExistsRequest;
-import in.koreatech.koin.domain.user.dto.OwnerLoginRequest;
-import in.koreatech.koin.domain.user.dto.OwnerLoginResponse;
+import in.koreatech.koin.domain.owner.dto.OwnerLoginRequest;
+import in.koreatech.koin.domain.owner.dto.OwnerLoginResponse;
 import in.koreatech.koin.domain.user.dto.StudentLoginRequest;
 import in.koreatech.koin.domain.user.dto.StudentLoginResponse;
 import in.koreatech.koin.domain.user.dto.StudentRegisterRequest;
@@ -40,7 +38,6 @@ import in.koreatech.koin.domain.user.dto.UserPasswordChangeRequest;
 import in.koreatech.koin.domain.user.dto.UserPasswordCheckRequest;
 import in.koreatech.koin.domain.user.dto.UserTokenRefreshRequest;
 import in.koreatech.koin.domain.user.dto.UserTokenRefreshResponse;
-import in.koreatech.koin.domain.user.model.UserType;
 import in.koreatech.koin.domain.user.service.StudentService;
 import in.koreatech.koin.domain.user.service.UserService;
 import in.koreatech.koin.global.auth.Auth;
@@ -94,25 +91,7 @@ public class UserController implements UserApi {
     public ResponseEntity<StudentLoginResponse> studentLogin(
         @RequestBody @Valid StudentLoginRequest request
     ) {
-        StudentLoginResponse response = userService.studentLogin(request);
-        return ResponseEntity.created(URI.create("/"))
-            .body(response);
-    }
-
-    @PostMapping("/owner/login")
-    public ResponseEntity<OwnerLoginResponse> ownerLogin(
-        @RequestBody @Valid OwnerLoginRequest request
-    ) {
-        OwnerLoginResponse response = userService.ownerLogin(request);
-        return ResponseEntity.created(URI.create("/"))
-            .body(response);
-    }
-
-    @PostMapping("/coop/login")
-    public ResponseEntity<CoopLoginResponse> coopLogin(
-        @RequestBody @Valid CoopLoginRequest request
-    ) {
-        CoopLoginResponse response = userService.coopLogin(request);
+        StudentLoginResponse response = studentService.studentLogin(request);
         return ResponseEntity.created(URI.create("/"))
             .body(response);
     }
